refactor(#42): 이미지 선택 훅으로 리팩토링
